# src/main/resources/META-INF/resources/openapi.yml
openapi: 3.0.0
info:
  title: Instant-crud documentation
  description: My version of the instant crud. Idea based of https://crudcrud.com/. <br> This is intended to be a fast and easy way to have a crud service for web development.
  version: 1.0.0
paths:
...
  /crud/{topic}:
    post:
      summary: Save data to a topic.
      tags:
        - Instant-crud
      description: Save data to a topic related to a api key and to the origin value.
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
      responses:
        '200':
          description: The data used to the topic in JSON format.
        '500':
          description: Returns a error that the data wasn't added.
    get:
      summary: Get all data from the a specific topic.
      tags:
        - Instant-crud
      description: Get all data from the a specific topic.
      responses:
        '200':
          description: Array with json object with the diferent data.
        '404':
          description: Empty array.
...